fix float number

+++ internal/dto/local_weather_dto.go
@@ -2,7 +2,7 @@ package dto
 
 type LocalWeatherDto struct {
 	Locale string
-	TempC  float32
-	TempF  float32
-	TempK  float32
+	TempC  float64
+	TempF  float64
+	TempK  float64
 }
